Split off ConstraintHandling into separate trait.
